+++ lib/ui/home/HomeCampusLinksWidget.dart
@@ -149,7 +149,10 @@ class _HomeCampusLinksWidgetState extends State<HomeCampusLinksWidget> implement
       // Build a default set of favorites
       List<String>? fullContent = JsonUtils.listStringsValue(FlexUI().contentSourceEntry('home.campus_links'));
       if (fullContent != null) {
-        Auth2().prefs?.setFavorites(HomeFavorite.favoriteKeyName(category: widget.favoriteId), favorites = LinkedHashSet<String>.from(fullContent.reversed));
+        favorites = LinkedHashSet<String>.from(fullContent.reversed);
+        Future.delayed(Duration(), () {
+          Auth2().prefs?.setFavorites(HomeFavorite.favoriteKeyName(category: widget.favoriteId), favorites);
+        });
       }
     }