+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Fixed section toggle from Browse panel [#1798](https://github.com/rokwire/illinois-app/issues/1798).
+- Fixed typo in "iDegrees New Student Checklist" from Browse / Academics [#1800](https://github.com/rokwire/illinois-app/issues/1800).
+
 ## [4.0.34] - 2022-06-21
 ### Added
 - Update GIES Step 2 [#1749](https://github.com/rokwire/illinois-app/issues/1749).